feat: support inline set value of select
Previously need to declare the select as a variable,
and set the value separately.
Otherwise he value is discarded before the options are add appended.

Currently set the value of select after all options are appended.
This enable inline setting the value in declarative way.
(and using expression, not requiring to use statement)